split applepay utils functions in different files
