--- packages/lib/src/components/ApplePay/config.ts
@@ -1,1 +1,0 @@
-export const APPLEPAY_SESSION_ENDPOINT = 'v1/applePay/sessions';
