+++ packages/lib/src/components/ApplePay/defaultProps.ts
@@ -1,4 +1,5 @@
 import { ApplePayConfiguration } from './types';
+import { inIframe } from '../../utils/inIframe';
 
 const defaultProps: ApplePayConfiguration = {
     isExpress: false,
@@ -9,7 +10,7 @@ const defaultProps: ApplePayConfiguration = {
     supportedNetworks: ['amex', 'discover', 'masterCard', 'visa'],
     buttonType: 'plain',
     buttonColor: 'black',
-    renderApplePayCodeAs: 'modal',
+    renderApplePayCodeAs: inIframe() ? 'window' : 'modal',
     onClick: resolve => resolve()
 };
