+++ packages/lib/src/components/ApplePay/utils/map-adyen-brands-to-applepay-brands.ts
@@ -1,0 +1,24 @@
+export function mapBrands(brands) {
+    const brandMapping = {
+        mc: 'masterCard',
+        amex: 'amex',
+        visa: 'visa',
+        elodebit: 'elo',
+        elo: 'elo',
+        interac: 'interac',
+        discover: 'discover',
+        jcb: 'jcb',
+        electron: 'electron',
+        maestro: 'maestro',
+        girocard: 'girocard',
+        cartebancaire: 'cartesBancaires',
+        eftpos_australia: 'eftpos'
+    };
+
+    return brands.reduce((accumulator, item) => {
+        if (!!brandMapping[item] && !accumulator.includes(brandMapping[item])) {
+            accumulator.push(brandMapping[item]);
+        }
+        return accumulator;
+    }, []);
+}
