+++ packages/lib/src/components/ApplePay/utils/payment-request.ts
@@ -1,5 +1,5 @@
-import { getDecimalAmount } from '../../utils/amount-util';
-import { PaymentAmount } from '../../types/global-types';
+import { getDecimalAmount } from '../../../utils/amount-util';
+import { PaymentAmount } from '../../../types/global-types';
 
 const formatAmount = (amount: PaymentAmount) => String(getDecimalAmount(amount.value, amount.currency));
