+++ packages/lib/src/components/ApplePay/utils/resolve-supported-version.ts
@@ -1,0 +1,12 @@
+export function resolveSupportedVersion(latestVersion: number): number | null {
+    try {
+        for (let v = latestVersion; v > 0; v--) {
+            if (ApplePaySession?.supportsVersion(v)) {
+                return v;
+            }
+        }
+    } catch (error) {
+        console.warn(error);
+    }
+    return null;
+}
