+++ packages/lib/src/utils/inIframe.ts
@@ -1,0 +1,3 @@
+const inIframe = () => window.self !== window.top;
+
+export { inIframe };
